Fixed issues with decimal formatting and pluralization rules

+++ engine/scripting/dotnet/src/RetroEngine.Portable/Localization/Cultures/PluralRules.cs
@@ -53,10 +53,10 @@ internal sealed partial class PluralRules : IDisposable
     private static partial void NativeClose(IntPtr rules);
 
     [LibraryImport(NativeLibraries.RetroCore, EntryPoint = "retro_plural_rules_select_int32")]
-    private static partial int NativeSelect(IntPtr rules, double number, Span<char> keyword, int capacity);
+    private static partial int NativeSelect(IntPtr rules, int number, Span<char> keyword, int capacity);
 
     [LibraryImport(NativeLibraries.RetroCore, EntryPoint = "retro_plural_rules_select_float64")]
-    private static partial int NativeSelect(IntPtr rules, int number, Span<char> keyword, int capacity);
+    private static partial int NativeSelect(IntPtr rules, double number, Span<char> keyword, int capacity);
 
     private void ReleaseUnmanagedResources()
     {
